Major code refactoring (api/detector/wmts)
